fixed readUInt, fixed induced error where the incorrect behavior was relied upon, added tests.
